+++ src/main/java/dinkplugin/notifiers/KillCountNotifier.java
@@ -1,18 +1,15 @@
 package dinkplugin.notifiers;
 
-import dinkplugin.message.Embed;
+import dinkplugin.message.NotificationBody;
+import dinkplugin.message.NotificationType;
 import dinkplugin.message.templating.Replacements;
 import dinkplugin.message.templating.Template;
-import dinkplugin.util.ItemUtils;
+import dinkplugin.notifiers.data.BossNotificationData;
 import dinkplugin.util.KillCountService;
 import dinkplugin.util.TimeUtils;
 import dinkplugin.util.Utils;
-import dinkplugin.message.NotificationBody;
-import dinkplugin.message.NotificationType;
-import dinkplugin.notifiers.data.BossNotificationData;
 import lombok.extern.slf4j.Slf4j;
-import net.runelite.api.NPC;
-import net.runelite.api.NpcID;
+import net.runelite.api.Client;
 import net.runelite.api.Varbits;
 import net.runelite.api.annotations.Varbit;
 import net.runelite.api.events.WidgetLoaded;
@@ -25,9 +22,6 @@ import org.jetbrains.annotations.VisibleForTesting;
 
 import javax.inject.Singleton;
 import java.time.Duration;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
@@ -79,7 +73,7 @@ public class KillCountNotifier extends BaseNotifier {
 
     public void onGameMessage(String message) {
         if (isEnabled())
-            parse(message).ifPresent(this::updateData);
+            parse(client, message).ifPresent(this::updateData);
     }
 
     public void onFriendsChatNotification(String message) {
@@ -98,7 +92,7 @@ public class KillCountNotifier extends BaseNotifier {
             // https://oldschool.runescape.wiki/w/Barbarian_Assault/Rewards#Earning_Honour_points
             if (widget != null && widget.getText().contains("80 ") && widget.getText().contains("5 ")) {
                 int gambleCount = client.getVarbitValue(Varbits.BA_GC);
-                this.data.set(new BossNotificationData(BA_BOSS_NAME, gambleCount, "The Queen is dead!", null, null));
+                this.data.set(new BossNotificationData(BA_BOSS_NAME, gambleCount, "The Queen is dead!", null, null, null));
             }
         }
     }
@@ -106,7 +100,7 @@ public class KillCountNotifier extends BaseNotifier {
     public void onTick() {
         BossNotificationData data = this.data.get();
         if (data != null) {
-            if (data.getBoss() != null) {
+            if (data.getBoss() != null && data.getCount() != null) {
                 // ensure notifier was not disabled during bad ticks wait period
                 if (isEnabled()) {
                     // once boss name has arrived, we notify at tick end (even if duration hasn't arrived)
@@ -144,34 +138,13 @@ public class KillCountNotifier extends BaseNotifier {
             .replacement("%TIME%", Replacements.ofText(time))
             .build();
 
-        // Prepare body
-        NotificationBody.NotificationBodyBuilder<BossNotificationData> body =
-            NotificationBody.<BossNotificationData>builder()
-                .text(content)
-                .extra(data)
-                .playerName(player)
-                .type(NotificationType.KILL_COUNT);
-
-        // Add embed if not screenshotting
-        boolean screenshot = config.killCountSendImage();
-        if (!screenshot && config.discordRichEmbeds()) {
-            if (ba) {
-                body.embeds(Collections.singletonList(Embed.ofImage(ItemUtils.getNpcImageUrl(NpcID.PENANCE_QUEEN))));
-            } else {
-                Arrays.stream(client.getCachedNPCs())
-                    .filter(Objects::nonNull)
-                    .filter(npc -> data.getBoss().equalsIgnoreCase(npc.getName()))
-                    .findAny()
-                    .map(NPC::getId)
-                    .map(ItemUtils::getNpcImageUrl)
-                    .map(Embed::ofImage)
-                    .map(Collections::singletonList)
-                    .ifPresent(body::embeds);
-            }
-        }
-
         // Call webhook
-        createMessage(screenshot, body.build());
+        createMessage(config.killCountSendImage(), NotificationBody.builder()
+            .text(content)
+            .extra(data)
+            .playerName(player)
+            .type(NotificationType.KILL_COUNT)
+            .build());
     }
 
     private boolean checkKillInterval(int killCount, @Nullable Boolean pb) {
@@ -193,23 +166,31 @@ public class KillCountNotifier extends BaseNotifier {
                 // Boss data and timing are sent in separate messages
                 // where the order of the messages differs depending on the boss.
                 // Here, we update data without setting any not-null values back to null.
+                String boss = defaultIfNull(updated.getBoss(), old.getBoss());
+                boolean tob = boss != null && boss.startsWith("Theatre of Blood"); // prefer challenge time message that comes first: https://github.com/pajlads/DinkPlugin/issues/585
                 return new BossNotificationData(
-                    defaultIfNull(updated.getBoss(), old.getBoss()),
+                    boss,
                     defaultIfNull(updated.getCount(), old.getCount()),
                     defaultIfNull(updated.getGameMessage(), old.getGameMessage()),
-                    defaultIfNull(updated.getTime(), old.getTime()),
-                    defaultIfNull(updated.isPersonalBest(), old.isPersonalBest())
+                    updated.getTime() == null || (tob && old.getTime() != null) ? old.getTime() : updated.getTime(),
+                    updated.isPersonalBest() == null || (tob && old.isPersonalBest() != null) ? old.isPersonalBest() : updated.isPersonalBest(),
+                    defaultIfNull(updated.getParty(), old.getParty())
                 );
             }
         });
     }
 
-    private static Optional<BossNotificationData> parse(String message) {
+    private static Optional<BossNotificationData> parse(Client client, String message) {
         if (message.startsWith("Preparation")) return Optional.empty();
         Optional<Pair<String, Integer>> boss = parseBoss(message);
         if (boss.isPresent())
-            return boss.map(pair -> new BossNotificationData(pair.getLeft(), pair.getRight(), message, null, null));
-        return parseTime(message).map(t -> new BossNotificationData(null, null, null, t.getLeft(), t.getRight()));
+            return boss.map(pair -> new BossNotificationData(pair.getLeft(), pair.getRight(), message, null, null, Utils.getBossParty(client, pair.getLeft())));
+
+        // TOB reports final wave duration before challenge time in the same message; skip to the part we care about
+        int tobIndex = message.startsWith("Wave") ? message.indexOf(KillCountService.TOB) : -1;
+        String msg = tobIndex < 0 ? message : message.substring(tobIndex);
+
+        return parseTime(msg).map(t -> new BossNotificationData(tobIndex < 0 ? null : KillCountService.TOB, null, null, t.getLeft(), t.getRight(), null));
     }
 
     private static Optional<Pair<Duration, Boolean>> parseTime(String message) {
@@ -278,10 +259,10 @@ public class KillCountNotifier extends BaseNotifier {
 
         int modeSeparator = boss.lastIndexOf(':');
         String raid = modeSeparator > 0 ? boss.substring(0, modeSeparator) : boss;
-        if (raid.equalsIgnoreCase("Theatre of Blood")
-            || raid.equalsIgnoreCase("Tombs of Amascut")
-            || raid.equalsIgnoreCase("Chambers of Xeric")
-            || raid.equalsIgnoreCase("Chambers of Xeric Challenge Mode"))
+        if (raid.equalsIgnoreCase(KillCountService.TOB)
+            || raid.equalsIgnoreCase(KillCountService.TOA)
+            || raid.equalsIgnoreCase(KillCountService.COX)
+            || raid.equalsIgnoreCase(KillCountService.COX + " Challenge Mode"))
             return boss;
 
         return null;